Implemented client side send operation
Implemented send operation for client side, send operation was
introduced with the LwM2M 1.1 standard.
This PR is based on work from @sbertin-telular, specifically on
send branch: https://github.com/sbertin-telular/wakaama/tree/send
In this PR a bug that caused a memory leak has been fixed:
https://github.com/sbertin-telular/wakaama/pull/1

+++ examples/lightclient/lightclient.c
@@ -202,7 +202,7 @@ void print_usage(void)
     fprintf(stdout, "  -l PORT\tSet the local UDP port of the Client. Default: 56830\r\n");
     fprintf(stdout, "  -4\t\tUse IPv4 connection. Default: IPv6 connection\r\n");
     fprintf(stdout, "  -S BYTES\tCoAP block size. Options: 16, 32, 64, 128, 256, 512, 1024. Default: %" PRIu16 "\r\n",
-            LWM2M_COAP_DEFAULT_BLOCK_SIZE);
+            (uint16_t)LWM2M_COAP_DEFAULT_BLOCK_SIZE);
     fprintf(stdout, "\r\n");
 }
